shaped builder rewritten

+++ shared/src/main/scala/org/denigma/semweb/messages/Read.scala
@@ -9,13 +9,13 @@ object Read {
   
   trait ReadMessage
 
-  case class Query(query:String,offset:Long = 0, limit:Long = Long.MaxValue,rewrite:Boolean = false) extends Paginated with ReadMessage
+  case class Query(query:String,offset:Int = 0, limit:Int = Int.MaxValue,rewrite:Boolean = false) extends Paginated with ReadMessage
 
-  case class Select(query:String,offset:Long = 0, limit:Long = Long.MaxValue,rewrite:Boolean = false) extends Paginated with ReadMessage
+  case class Select(query:String,offset:Int = 0, limit:Int = Int.MaxValue,rewrite:Boolean = false) extends Paginated with ReadMessage
 
-  case class Bind(query:String,binds:Map[String,String],offset:Long = 0, limit:Long = Long.MaxValue) extends Paginated with ReadMessage
+  case class Bind(query:String,binds:Map[String,String],offset:Int = 0, limit:Int = Int.MaxValue) extends Paginated with ReadMessage
 
-  case class Search(query:String,searches:Map[String,String],binds:Map[String,String] = Map.empty,offset:Long = 0, limit:Long = Long.MaxValue) extends Paginated with ReadMessage
+  case class Search(query:String,searches:Map[String,String],binds:Map[String,String] = Map.empty,offset:Int = 0, limit:Int = Int.MaxValue) extends Paginated with ReadMessage
 
   case class Question(query:String) extends QueryLike with ReadMessage
 
@@ -29,8 +29,8 @@ object Read {
 
 
 trait Paginated extends QueryLike{
-  def offset:Long
-  def limit:Long
-  def isPaginated = offset>0 || (limit != Long.MaxValue && limit>0)
+  def offset:Int
+  def limit:Int
+  def isPaginated = offset>0 || (limit != Int.MaxValue && limit>0)
 
 }
