+++ shared/src/main/scala/org/denigma/semweb/messages/SimpleRead.scala
@@ -13,11 +13,11 @@ object SimpleRead {
   trait SimpleQuery extends QueryLike
 
 
-  case class Select(query:String,offset:Long = 0, limit:Long = Long.MaxValue,rewrite:Boolean = false) extends Paginated with SimpleQuery
+  case class Select(query:String,offset:Int = 0, limit:Int = Int.MaxValue,rewrite:Boolean = false) extends Paginated with SimpleQuery
 
-  case class Bind(query:String,binds:Map[String,String],offset:Long = 0, limit:Long = Long.MaxValue) extends Paginated  with SimpleQuery
+  case class Bind(query:String,binds:Map[String,String],offset:Int = 0, limit:Int = Int.MaxValue) extends Paginated  with SimpleQuery
 
-  case class Search(query:String,searches:Map[String,String],binds:Map[String,String] = Map.empty,offset:Long = 0, limit:Long = Long.MaxValue,params:Map[String,String]) extends Paginated with SimpleQuery
+  case class Search(query:String,searches:Map[String,String],binds:Map[String,String] = Map.empty,offset:Int = 0, limit:Int = Int.MaxValue,params:Map[String,String]) extends Paginated with SimpleQuery
 
   case class Question(query:String) extends QueryLike with SimpleQuery
